agent/requirements.txt
# agent/requirements.txt

# ── Core API ────────────────────────────────────────────────────────────────
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
pydantic>=2.7.4
httpx>=0.27.0

# ── LangChain ecosystem ─────────────────────────────────────────────────────
langchain>=0.3.0,<0.4.0
langchain-core>=0.3.0,<0.4.0
langchain-community>=0.3.0,<0.4.0

# ── LLM Provider 1: Ollama (local in-cluster) ───────────────────────────────
ollama>=0.6.0

# ── LLM Provider 2: Vertex AI / Gemini (GCP service-account JSON) ───────────
google-cloud-aiplatform>=1.140.0
google-auth>=2.49.0
google-auth-httplib2>=0.3.0

# ── LLM Provider 3: AWS Bedrock (IAM role / IRSA - no static keys needed) ───
boto3>=1.34.0

# ── Kubernetes (in-cluster auth) ─────────────────────────────────────────────
kubernetes>=30.0.0

# ── Elasticsearch / Kibana ───────────────────────────────────────────────────
elasticsearch>=8.13.0

# ── HTTP requests (Jenkins, Artifactory, Nginx, Kibana REST) ─────────────────
requests>=2.32.0
urllib3>=2.2.0

# ── Utilities ────────────────────────────────────────────────────────────────
python-dotenv>=1.2.0
tenacity>=8.3.0
cachetools>=5.5.0

# ── Observability ────────────────────────────────────────────────────────────
prometheus-client>=0.21.0
